+++ src/decomp_workbench/fingerprints/schedule.c
@@ -1,3 +1,7 @@
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 volatile int dkwb_fp_left;
 volatile int dkwb_fp_right;
 
@@ -7,3 +11,7 @@ int dkwb_fp_schedule(int value)
     int right = dkwb_fp_right + 10;
     return (left * value) + right;
 }
+
+#ifdef __cplusplus
+}
+#endif